Attribut lz_code als Integer in Pub-DB

+++ alw_zonengrenzen_import/alw_zonengrenzen_pub.sql
@@ -2,7 +2,7 @@ SELECT
     lz_flaeche.t_id,
     lz_flaeche.flaeche,
     erstellungsdatum,
-    katalogtyp.lz_code,
+    katalogtyp.lz_code::int,
     katalogtyp.typ_de AS typ
 FROM
     alw_zonengrenzen.zonengrenzen_lz_flaeche AS lz_flaeche
